Temperature Sensor Full Implementation
Can now read from the temperature sensor and also calculate the current speed of sound. 
Temperature sensor uses the PA12 pin, renamed "DHT22", which acts both a input/output GPIO.

The timing is very finicky for the DHT22_Start(), DHT22_Check_Response(), and DHT22_Read() functions due to the output nature of the DHT22 sensor (which relies on specific timings to change its state). If they are not working, try messing around with the usDelay() amount.